% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd_slc.R
\name{gcd_slc}
\alias{gcd_slc}
\title{Calculate distance based on spherical law of cosines}
\usage{
gcd_slc(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{starting longitude point}

\item{lat1}{starting latitude point}

\item{long2}{ending longitude point}

\item{lat2}{ending latitude point}
}
\value{
Distance in km
}
\description{
Calculate distance based on spherical law of cosines
}

